fix todo:Move function readinessCheck to util
